Support for post back mode
When CxFlow is running in post-back mode poIing will be turned off and CxFlow will be switched to an event driven mode. CxFlow will wait for the manager to post a response when the scan is complete. This is an optional mode of operation enabled by setting enable-post-action-monitor to true in the application config file, the default is to use the traditional poling mode.