Set project custom fields when creating/updating project for scan
Just as we already support updating the preset, engine configuration
and file and directory exclusions when creating a project, with this
change we set the project's custom fields. This wil be done if the
project is created or, if the project already exists, the settings
override flag has been set.