Handle race condition when creating a project
It is possible that two CxFlow processes, running in parallel, will
both try to create the same project. One process will succeed and the
other will receive a 400 (Bad Request) response indicating that the
project already exists. In the latter case, this change causes CxFlow
to retrieve the existing project's identifier and return it.